Create SQL_Activity_1.txt
Objective: To identify database tables and columns.
For the Associates: Associates to finish this activity before proceeding to the next activity. It is enough to identify the tables and column names, no need to create.
Scenario: ABC college wants to develop a University Management System (UMS) to store information on students who join their college. The database should contain
• Student’s personal information as well as student’s academic details.
• It should store information on subjects taught at the university in various departments.
• It should also store marks obtained by each student in each semester subject wise as well as the GPA for each semester.
GPA is a rating calculated in a scale of ten considering the individual subject marks obtained and subjects weightage % in a semester

Problem Statement: Using the following and identify the tables and columns needed for University Management System. • Student information should have registration_number, name, branch , contact #, DOB, Date of joining, Address, Email id. • Information on subjects like subject code, subject name and weightage for calculating GPA. • Students marks scored in each subject , semester wise. • Finally overall result of the student comprising GPA scored for a semester, scholarship eligibility. Associates should use the above case and identify the tables and columns for building the UMS system. IMPORTANT NOTE: The number of subjects varies from semester to semester and the university also changes the number of subjects or swaps the subjects in a particular semester. The database table design should be in such a way that any new subject additions or subject removal should not impact the database design that is I should not add columns or alter tables. The design should be flexible for changes.